+++ src/Swasey/Model/Internal/OperationDefinition.cs
@@ -161,5 +161,6 @@ namespace Swasey.Model
             return this;
         }
 
+
     }
 }